+++ src/templates/base.html
@@ -1,3 +1,5 @@
+{% from "macros.html" import alert %}
+
 <!DOCTYPE html>
 <html>
 <head>
@@ -14,6 +16,7 @@
 </head>
 
 <body>
+  {% if error %}{{alert(error)}}{% endif %}
   <main class="main">{% block main %}{% endblock %}</main>
   {% block script %}{% endblock %}
 </body>
