+++ src/templates/user/login.html
@@ -11,10 +11,10 @@
     <h4 class="card-header">Login</h4>
     <div class="card-body">
       <form  action="/login" method="POST">
-        <div class="form-group">
+        <div class="form-group mandatory">
           <input name="username" type="text" class="form-control" placeholder="Usuário">
         </div>
-        <div class="form-group">
+        <div class="form-group mandatory">
           <input name="password" type="password" class="form-control" placeholder="Senha">
         </div>
         <button class="btn btn-success btn-block">Entrar</button>
